v0.1.0a-Restarting
-Made the game restartable(memory leak needs fixing).
-Removed the possibility of the IP containing letters.
-Fixed end of InputBox sometimes getting garbage value.
-Server shouldn't show MessageBoxes anymore.
-Cleaned hardcoded values regarding 3D picking in the Menu.

+++ TorpedoJatekClient/Source/Backend/EventHandler.cpp
@@ -99,7 +99,7 @@ void EventHandler::ApplySoundDistEffect(int channel_num, float sound_pos_x, floa
 	float len = glm::length(glm::vec3(camEyeX - sound_pos_x, 0
 		, camEyeZ - sound_pos_z));
 	float dis = (len > soundScanDistance ? soundScanDistance : len) * 255 / soundScanDistance;
-	Mix_SetDistance(channel_num, static_cast<int>(dis));
+	Mix_SetDistance(channel_num, static_cast<Uint8>(dis));
 }
 
 //Muteolja vagy reseteli az audio hangerõt - Zene és channelek
